Cleaned up links
I changed up remove some of the references to the slider technology
team 7 was using in their source code.

I couldn’t check if it was having a significant performance impact, so
it is left in as a comment if anyone wants to implement the old
slideshow solution at a later time.